update request handler
