+++ hodor-actuator/hodor-actuator-common/src/main/java/org/dromara/hodor/actuator/common/action/AbstractExecuteAction.java
@@ -38,8 +38,6 @@ public abstract class AbstractExecuteAction extends AbstractAction<JobExecuteReq
 
     private final HodorProperties properties;
 
-    private final ExecutorManager executorManager;
-
     private final JobExecutionPersistence jobExecutionPersistence;
 
     private final JobLoggerManager jobLoggerManager;
@@ -48,12 +46,10 @@ public abstract class AbstractExecuteAction extends AbstractAction<JobExecuteReq
 
     public AbstractExecuteAction(final RequestContext context,
                                  final HodorProperties properties,
-                                 final ExecutorManager executorManager,
                                  final JobExecutionPersistence jobExecutionPersistence,
                                  final RequestHandleManager requestHandleManager) {
         super(context, requestHandleManager);
         this.properties = properties;
-        this.executorManager = executorManager;
         this.jobExecutionPersistence = jobExecutionPersistence;
         this.jobLoggerManager = JobLoggerManager.getInstance();
         this.stopwatch = Stopwatch.create();
@@ -135,7 +131,7 @@ public abstract class AbstractExecuteAction extends AbstractAction<JobExecuteReq
     @Override
     public void afterProcess() {
         jobLogger.info("job execution finished.");
-        executorManager.removeExecutableNode(requestId);
+        getRequestHandleManager().removeExecutableNode(requestId);
         jobLoggerManager.stopJobLogger(loggerName);
     }
